Add SMART helpers
